Refactor agility course plugins into subpackages
Moved GnomeStrongholdCoursePlugin and DraynorRooftopCoursePlugin into 'regularcourses' and 'rooftopcourses' subpackages, respectively. Updated package declarations and imports to reflect new structure, improving code organization and maintainability.
